+++ sentry/src/main/java/io/sentry/NoOpSentryClient.java
@@ -24,7 +24,7 @@ final class NoOpSentryClient implements ISentryClient {
 
   @Override
   public @NotNull SentryId captureEvent(
-      @NotNull SentryEvent event, @Nullable Scope scope, @Nullable Hint hint) {
+      @NotNull SentryEvent event, @Nullable IScope scope, @Nullable Hint hint) {
     return SentryId.EMPTY_ID;
   }
 
@@ -49,7 +49,7 @@ final class NoOpSentryClient implements ISentryClient {
   public @NotNull SentryId captureTransaction(
       @NotNull SentryTransaction transaction,
       @Nullable TraceContext traceContext,
-      @Nullable Scope scope,
+      @Nullable IScope scope,
       @Nullable Hint hint,
       @Nullable ProfilingTraceData profilingTraceData) {
     return SentryId.EMPTY_ID;
@@ -58,7 +58,7 @@ final class NoOpSentryClient implements ISentryClient {
   @Override
   @ApiStatus.Experimental
   public @NotNull SentryId captureCheckIn(
-      @NotNull CheckIn checkIn, @Nullable Scope scope, @Nullable Hint hint) {
+      @NotNull CheckIn checkIn, @Nullable IScope scope, @Nullable Hint hint) {
     return SentryId.EMPTY_ID;
   }