+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -90,7 +90,7 @@ public final class SentryClient implements ISentryClient {
 
   @Override
   public @NotNull SentryId captureEvent(
-      @NotNull SentryEvent event, final @Nullable Scope scope, @Nullable Hint hint) {
+      @NotNull SentryEvent event, final @Nullable IScope scope, @Nullable Hint hint) {
     Objects.requireNonNull(event, "SentryEvent is required.");
 
     if (hint == null) {
@@ -246,7 +246,7 @@ public final class SentryClient implements ISentryClient {
     return sentryId;
   }
 
-  private void addScopeAttachmentsToHint(@Nullable Scope scope, @NotNull Hint hint) {
+  private void addScopeAttachmentsToHint(@Nullable IScope scope, @NotNull Hint hint) {
     if (scope != null) {
       hint.addAttachments(scope.getAttachments());
     }
@@ -495,7 +495,7 @@ public final class SentryClient implements ISentryClient {
   @TestOnly
   @Nullable
   Session updateSessionData(
-      final @NotNull SentryEvent event, final @NotNull Hint hint, final @Nullable Scope scope) {
+      final @NotNull SentryEvent event, final @NotNull Hint hint, final @Nullable IScope scope) {
     Session clonedSession = null;
 
     if (HintUtils.shouldApplyScopeData(hint)) {
@@ -599,7 +599,7 @@ public final class SentryClient implements ISentryClient {
   public @NotNull SentryId captureTransaction(
       @NotNull SentryTransaction transaction,
       @Nullable TraceContext traceContext,
-      final @Nullable Scope scope,
+      final @Nullable IScope scope,
       @Nullable Hint hint,
       final @Nullable ProfilingTraceData profilingTraceData) {
     Objects.requireNonNull(transaction, "Transaction is required.");
@@ -681,7 +681,7 @@ public final class SentryClient implements ISentryClient {
   @Override
   @ApiStatus.Experimental
   public @NotNull SentryId captureCheckIn(
-      @NotNull CheckIn checkIn, final @Nullable Scope scope, @Nullable Hint hint) {
+      @NotNull CheckIn checkIn, final @Nullable IScope scope, @Nullable Hint hint) {
     if (hint == null) {
       hint = new Hint();
     }
@@ -758,7 +758,7 @@ public final class SentryClient implements ISentryClient {
   }
 
   private @Nullable SentryEvent applyScope(
-      @NotNull SentryEvent event, final @Nullable Scope scope, final @NotNull Hint hint) {
+      @NotNull SentryEvent event, final @Nullable IScope scope, final @NotNull Hint hint) {
     if (scope != null) {
       applyScope(event, scope);
 
@@ -789,7 +789,7 @@ public final class SentryClient implements ISentryClient {
     return event;
   }
 
-  private @NotNull CheckIn applyScope(@NotNull CheckIn checkIn, final @Nullable Scope scope) {
+  private @NotNull CheckIn applyScope(@NotNull CheckIn checkIn, final @Nullable IScope scope) {
     if (scope != null) {
       // Set trace data from active span to connect events with transactions
       final ISpan span = scope.getSpan();
@@ -807,7 +807,7 @@ public final class SentryClient implements ISentryClient {
   }
 
   private <T extends SentryBaseEvent> @NotNull T applyScope(
-      final @NotNull T sentryBaseEvent, final @Nullable Scope scope) {
+      final @NotNull T sentryBaseEvent, final @Nullable IScope scope) {
     if (scope != null) {
       if (sentryBaseEvent.getRequest() == null) {
         sentryBaseEvent.setRequest(scope.getRequest());